Made a small edit to get it to work with python 3.7.9
it might be because the original author was using an older version.  But problem is that in the list2dict, for some proxy types there may not be an x[7].  Perhaps its due to the implementation of the underlying modules like urllib and re changed after python 2.8 since the error from pip, pip3 is "ERROR: Package 'proxify' requires a different Python: 3.7.9 not in '<=2.8,>=2'"

Hopefully i can just rewrap this file and upload it to the pip database as proxify379 or something like that.  If people want to use it they can, but in general i'm just making this for one of my other tools to easily fetch proxies and use them right away for some concurrent processing tasks.